swap ptx.py
